Adjust dependencies
