+++ notifications/src/api/v1/notify.py
@@ -3,12 +3,11 @@ from http import HTTPStatus
 from typing import Annotated
 
 from fastapi import APIRouter, Body, Depends, HTTPException
-from faststream.rabbit.fastapi import RabbitBroker
-from src.db.rabbitmq import default_queue, get_rabbitmq_broker
 from src.services.notifications_service import (
     NotificationsService,
     get_notifications_service,
 )
+from src.services.queue_service import QueueService, get_queue
 
 from . import schemas
 
@@ -20,7 +19,7 @@ logger = logging.getLogger(__name__)
 @router.post("/notify", summary="Отправляем нотификация", response_model=None)
 async def do_notify(
     notification: Annotated[schemas.CreateNotification, Body()],
-    rabbitmq: Annotated[RabbitBroker, Depends(get_rabbitmq_broker)],
+    queue: Annotated[QueueService, Depends(get_queue)],
     service: Annotated[NotificationsService, Depends(get_notifications_service)],
 ) -> schemas.NotificationSummary:
     try:
@@ -28,8 +27,7 @@ async def do_notify(
             notification.subject, notification.body, notification.distribution, notification.schedule
         )
         if model.last_sent:
-            status = await rabbitmq.publish({"m": {"notification_id": model.id}}, queue=default_queue)
-            logger.info(status)
+            await queue.publish_notification(model)
 
         return schemas.NotificationSummary.model_validate(model)
 
@@ -41,14 +39,13 @@ async def do_notify(
 async def notify_by_template(
     template_id: str,
     notification: Annotated[schemas.TemplateNotification, Body()],
-    rabbitmq: Annotated[RabbitBroker, Depends(get_rabbitmq_broker)],
+    queue: Annotated[QueueService, Depends(get_queue)],
     service: Annotated[NotificationsService, Depends(get_notifications_service)],
 ) -> schemas.NotificationSummary:
     try:
         model = await service.create_from_template(template_id, notification.distribution, notification.schedule)
         if model.last_sent:
-            status = await rabbitmq.publish({"m": {"notification_id": model.id}}, queue=default_queue)
-            logger.info(status)
+            await queue.publish_notification(model)
         return schemas.NotificationSummary.model_validate(model)
     except RuntimeError as err:
         raise HTTPException(status_code=HTTPStatus.BAD_REQUEST, detail=repr(err))
