+++ notifications/src/models/queue_payload.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+from .notification import Notification
+
+
+class QueuePayload(BaseModel):
+    message: Notification
