+++ notifications/src/services/notifications_service.py
@@ -1,6 +1,5 @@
 from datetime import UTC, datetime, timedelta
-from typing import Annotated, cast
-from uuid import UUID
+from typing import Annotated
 
 from bson import ObjectId
 from croniter import croniter
